+++ projects/komponentkartan/src/lib/controls/datepicker/datepicker.component.ts
@@ -36,21 +36,9 @@ export class DatepickerComponent implements OnChanges, AfterViewInit, OnDestroy,
   @ViewChild('headerLabel') headerLabel: ElementRef;
   @ViewChild('headerInput') headerInput: ElementRef;
   @ViewChild('readOnlyHeader') readOnlyHeader: ElementRef;
+  @ViewChild('headerInputDiv') headerInputDiv: ElementRef;
   @ViewChild('calendar') calendar: ElementRef;
   @ViewChildren(DatepickerItemComponent) items: QueryList<DatepickerItemComponent>;
-
-  @HostListener('window:resize', []) onWindowResize() {
-    // only change if 'default' values
-    if (window.innerWidth <= 1459) {
-      if (this.width === '142px') {
-        this.width = '130px';
-      }
-    } else {
-      if (this.width === '130px') {
-        this.width = '142px';
-      }
-    }
-  }
 
   headerLabelId = Guid.newGuid();
   label = '';
@@ -72,6 +60,22 @@ export class DatepickerComponent implements OnChanges, AfterViewInit, OnDestroy,
   private actualMaxDate: Date;
   private ngUnsubscribe = new Subject();
   private ngUnsubscribeItems = new Subject();
+  elementId: string;
+
+  @HostListener('window:resize', [])
+  onWindowResize() {
+    // only change if 'default' values
+    if (window.innerWidth <= 1459) {
+      if (this.width === '142px') {
+        this.width = '130px';
+      }
+    } else {
+      if (this.width === '130px') {
+        this.width = '142px';
+      }
+    }
+  }
+
 
   get errorActive() {
     if (!this.showValidation || this.disabled || this.readonly) {
@@ -90,6 +94,7 @@ export class DatepickerComponent implements OnChanges, AfterViewInit, OnDestroy,
   }
 
   constructor(@Inject(LOCALE_ID) private locale: string, @Optional() @Self() public formControl: NgControl) {
+    this.elementId = Math.random().toString();
     this.setMinZoomLevel();
 
     if (this.formControl != null) {
@@ -206,6 +211,13 @@ export class DatepickerComponent implements OnChanges, AfterViewInit, OnDestroy,
     this.collapse(false);
   }
 
+  onHeaderInputKeydown(event: KeyboardEvent ) {
+    // to be able to move backwards-tab, otherwise we will be stuck in an endless loop going to input in div 'allowtext'
+    if (event.key === 'Tab' && event.shiftKey) {
+      this.headerInputDiv.nativeElement.focus();
+    }
+  }
+
   onKeydown(event: KeyboardEvent) {
     if (this.disabled || this.readonly) {
       return;
@@ -213,8 +225,13 @@ export class DatepickerComponent implements OnChanges, AfterViewInit, OnDestroy,
 
     if (event.key === 'Escape' || event.key === 'Esc') {
       this.collapse();
-    } else if (event.key === 'Tab') {
-      this.collapse();
+    } else if (event.key === 'Tab' ) {
+      if (event.shiftKey) {
+        // to be able to move backwards-tab, otherwise we will be stuck in an endless loop going to input in div 'allowtext'
+        this.collapse(false);
+      } else {
+        this.collapse();
+      }
     } else if (event.key === 'Home') {
       event.preventDefault();
       this.items.first.focus();
@@ -478,6 +495,10 @@ export class DatepickerComponent implements OnChanges, AfterViewInit, OnDestroy,
   }
 
   private collapse(focusHeader = true) {
+    if (this.disabled) {
+      return;
+    }
+
     this.expanded = false;
 
     if (focusHeader) {
@@ -921,6 +942,15 @@ export class DatepickerComponent implements OnChanges, AfterViewInit, OnDestroy,
           }
         });
     });
+  }
+
+  getLabelFromId() {
+    // return window.document.getElementById(this.idForLabel)
+    let labels = document.getElementsByTagName('label');
+    for( var i = 0; i < labels.length; i++ ) {
+      if (labels[i].htmlFor == this.labelId)
+           return labels[i];
+   }
   }
 }