+++ projects/komponentkartan/src/lib/controls/grid/grid-header-column.component.ts
@@ -1,15 +1,12 @@
 import { Component, OnInit, Input, Output, EventEmitter, HostBinding, ViewChild, ElementRef } from '@angular/core';
+import { GridSortDirection } from '../sort-arrow/sort-arrow.component';
 
-export enum GridSortDirection {
-  None,
-  Ascending,
-  Descending
-}
+
 @Component({
   selector: 'vgr-grid-header-column',
   templateUrl: './grid-header-column.component.html'
 })
-export class GridHeaderColumnComponent implements OnInit {
+export class GridHeaderColumnComponent {
   @Input() @HostBinding('style.flex') width = 1;
   @Input() align = 'left';
   @Input() sortKey: string;
@@ -18,9 +15,6 @@ export class GridHeaderColumnComponent implements OnInit {
   @ViewChild('gridHeaderColumn', { read: ElementRef, static: false }) gridHeaderColumn: ElementRef;
 
   constructor() { }
-
-  ngOnInit() {
-  }
 
   changeSort() {
     if (this.sortKey) {