+++ projects/komponentkartan/src/lib/controls/menu/menu-item.component.ts
@@ -33,6 +33,12 @@ export class MenuItemComponent extends MenuItemBaseDirective implements AfterVie
 
             if (this.link) {
                 this.isInternalLink ? this.router.navigate([this.link]) : this.onExternalLink();
+              setTimeout(() => {
+                const id = document.getElementById('page-content-focus');
+                if (id) {
+                  this.renderer.selectRootElement('#page-content-focus', true).focus();
+                }
+              }, 100);
             } else if (this.action) {
                 this.onAction(event);
             }