+++ projects/komponentkartan/src/lib/controls/sort-arrow/sort-arrow.component.ts
@@ -1,20 +1,22 @@
-import { Component, OnInit, Input, OnChanges, HostBinding } from '@angular/core';
-import { GridSortDirection } from '../grid/grid-header-column.component';
+import { Component, Input, OnChanges, HostBinding } from '@angular/core';
 
+export enum GridSortDirection {
+  None,
+  Ascending,
+  Descending
+}
 @Component({
   selector: 'vgr-sort-arrow',
   templateUrl: './sort-arrow.component.html',
   styleUrls: ['./sort-arrow.component.scss']
 })
-export class SortArrowComponent implements OnInit, OnChanges {
+
+export class SortArrowComponent implements  OnChanges {
 
   @Input() sortDirection: GridSortDirection;
   @HostBinding('class') className;
 
   constructor() { }
-
-  ngOnInit(): void {
-  }
 
   ngOnChanges(changes) {
     const values = ['only-hover', 'permanent', 'permanent'];