+++ projects/komponentkartan/src/lib/controls/tab-button/tab-button-group.component.spec.ts
@@ -22,7 +22,6 @@ describe('TabButtonGroupComponent', () => {
   let component: TabButtonGroupComponent;
   let fixture: ComponentFixture<TestComponent>;
   let rootElement: DebugElement;
-
   let firstTabButton: DebugElement;
   let firstTabButtonOuterSpan: DebugElement;
   let firstTabButtonButton: DebugElement;