+++ projects/komponentkartan/src/lib/controls/tab-button/tab-button.component.spec.ts
@@ -29,7 +29,7 @@ describe('TabButtonComponent', () => {
   });
 
   describe('When button is disabled', () => {
-    let buttonElement
+    let buttonElement;
     beforeEach(() => {
       buttonElement = rootElement.query(By.css('button'));
       component.disabled = true;
@@ -51,7 +51,7 @@ describe('TabButtonComponent', () => {
   })
 
   describe('When disabled is false', () => {
-    let buttonElement
+    let buttonElement;
     beforeEach(() => {
       buttonElement = rootElement.query(By.css('button'));
       component.disabled = false;
@@ -73,7 +73,7 @@ describe('TabButtonComponent', () => {
   });
 
   describe('When button is already active and pressed again', () => {
-    let buttonElement
+    let buttonElement;
     let spy;
     beforeEach(() => {
       buttonElement = rootElement.query(By.css('button'));
@@ -88,7 +88,7 @@ describe('TabButtonComponent', () => {
 
       fixture.detectChanges();
       spy = spyOn(component.selectedChanged, 'emit').and.callThrough();
-    })
+    });
 
     it('button event is not propogated', () => {
       expect(spy).not.toHaveBeenCalled();
@@ -96,7 +96,7 @@ describe('TabButtonComponent', () => {
   });
 
   describe('WCAG Tests', () => {
-    let buttonElement
+    let buttonElement;
     beforeEach(() => {
       buttonElement = rootElement.query(By.css('button'));
     });