+++ projects/komponentkartan/src/lib/controls/tab-button/tab-button.component.ts
@@ -1,4 +1,4 @@
-import { AfterViewInit, Component, ElementRef, EventEmitter, Input, OnChanges, OnInit, Output, ViewChild } from '@angular/core';
+import { AfterViewInit, Component, ElementRef, EventEmitter, Input, OnChanges, OnInit, Output, Renderer2, ViewChild } from '@angular/core';
 import { Guid } from '../../utils/guid';
 
 @Component({
@@ -17,6 +17,7 @@ export class TabButtonComponent implements AfterViewInit, OnChanges {
   @Output() next = new EventEmitter();
   @Output() previous = new EventEmitter();
   @Output() home = new EventEmitter();
+  // eslint-disable-next-line @angular-eslint/no-output-native
   @Output() end = new EventEmitter();
   @Output() selectedChanged = new EventEmitter<string>();
 
@@ -27,7 +28,7 @@ export class TabButtonComponent implements AfterViewInit, OnChanges {
   ariaPressed = false;
   initialTabId: string;
 
-  constructor(private elementRef: ElementRef) {}
+  constructor(private elementRef: ElementRef, private renderer: Renderer2) {}
 
   ngOnChanges() {
     Promise.resolve(null).then(() => {