+++ projects/komponentkartan/src/lib/pipes/errorMessagePipe.ts
@@ -7,23 +7,23 @@ import { Pipe, PipeTransform } from '@angular/core';
 export class ErrorMessagePipe implements PipeTransform {
   currentMessage: string;
 
-  transform(message: any, errors: any, small = false) {
+  transform(message: any, errors: any, label: any, small = false) {
     if (!message) {
       return null;
     }
 
-    this.currentMessage = this.setErrorMessage(errors, message, small);
+    this.currentMessage = this.setErrorMessage(errors, message, label, small);
 
     return this.currentMessage;
   }
 
-  private setErrorMessage(errors: any, message: any, small: any): string {
+  private setErrorMessage(errors: any, message: any, label: any, small: any): string {
     if (typeof (message) === 'object') {
       for (const key in errors) {
         if (key === 'required' && small && !message[key]) {
           return 'Obligatoriskt';
         } else if (key === 'required' && !small && !message[key]) {
-          return 'Fältet är obligatoriskt';
+          return label ? label.innerText + ' är obligatoriskt' : 'Fältet är obligatoriskt';
         } else if (message) {
           return message[key];
         } else {